feat: Week1 GUI系リファクタリング完全実装
Phase2 Week1完了: 3大GUIファイルの完全分割

## 📊 分割成果
### gui_views.py: 539行 → 最大176行（67%削減）
- main_window.py: 49行（ウィンドウ管理）
- widgets.py: 163行（基本ウィジェット）
- controls.py: 176行（コントロール要素）
- dialogs.py: 93行（ダイアログ）
- main_view.py: 96行（統合ビュー）

### gui_controller.py: 397行 → 最大221行（44%削減）
- main_controller.py: 144行（統合制御・イベント管理）
- file_controller.py: 87行（ファイル操作制御）
- conversion_controller.py: 221行（変換・サンプル生成制御）
- gui_controller.py: 46行（統合インターフェース）

### gui_models.py: 297行 → 最大153行（48%削減）
- config_model.py: 153行（設定・変換状態管理）
- file_model.py: 53行（ファイル操作管理）
- state_model.py: 109行（ログ・アプリ状態管理）

## 🏗️ アーキテクチャ改善
- ✅ Single Responsibility Principle完全適用
- ✅ MVC分離強化（Model/View/Controller明確化）
- ✅ 循環インポート回避（TYPE_CHECKING活用）
- ✅ 後方互換性維持（レガシーラッパー）
- ✅ モジュール分離（機能別責任分担）

## 📈 技術的負債改善
- **37件 → 34件**（3件解消）
- GUI系300行制限違反：**完全解消**
- 保守性・テスタビリティ大幅向上

## ✅ 動作確認済み
- 新モジュール構造インポート成功
- レガシー互換性確認成功
- 静的メソッドテスト成功

## 🎯 次のステップ（Week2）
コアパーサー系リファクタリング開始予定：
- keyword_parser.py（521行）→ 戦略パターン適用
- list_parser.py（383行）→ 責任分離

健全で保守性の高いGUI層アーキテクチャを確立 🚀
